ReadMe.md
# Predictive Stock Market LSTM Model

Hello and welcome to this notebook. 

The following project was build in a python jupyter notebook. 
The project consists of using a dataset of stock market information from 85 of the
100 companies in 
the NASDAQ 100 list as of July 26th, 2021.

The structure of the project is as follows:
1. Python Imported Packages
2. Data Overview and Loading the Dataset 
3. Exploratory Data Analysis 
4. Data Cleansing/PreProcessing 
5. Model Preparation & Train-Test Split
6. Model Performance Metrics

This project consists of analzing stock market information and finding out useful information
and trends without the datasets, data cleaning using pandas and numpy amd building a LSTM (Long Short Term Memory)
model using our data to predict stock prices.

The dataset is included as a link in the notebook and explanations of the data cleansing
and data preparation process is also included.

Enjoy!



## Authors

- [@AlexMason](https://github.com/alexmason97)

  
## Python Packages Used in Notebook

- datetime
- keras
- matplotlib.pyplot
- numpy
- pandas
- pandas_datareader.data
- seaborn
- sklearn.preprocessing
- tensorflow
- time
- warnings


  
## License

[MIT](https://choosealicense.com/licenses/mit/)

MIT License

Copyright (c) [2021]

Permission is hereby granted, free of charge, to any person obtaining a copy
of this software and associated documentation files (the "Software"), to deal
in the Software without restriction, including without limitation the rights
to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
copies of the Software, and to permit persons to whom the Software is
furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all
copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
SOFTWARE.
## Lessons Learned

Q: What did you learn while building this project? 

A: Learned a lot about LSTM models and what goes into building an LSTM model to 
be able to 'look back' at data entries to be able to make predictions on future values.
Also learned quite a bit of tensorflow and model optimization as well

Q: What challenges did you face and how did you overcome them?

A: Many issues with data manipulation and specific function calls not working due to the datatypes
used in the dataset. Was able to overcome them or figure out work arounds that suited the needs of this project's 
appropriate completion.

  
## Project Checklist

- [x] Data Analysis
- [x] Data Cleansing
- [x] Data Visualization
- [x] Model Prediction